api/requirements.txt
attrs==17.4.0
Automat==0.6.0
blinker==1.4
chardet==3.0.4
click==6.7
gunicorn==20.1.0
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
mysql-connector-python==8.0.25
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==16.0.0
six==1.16.0
urllib3==1.22
